+++ tests/tibanna/unicorn/test_ec2_utils.py
@@ -1,4 +1,6 @@
-from tibanna.ec2_utils import update_config
+from tibanna.ec2_utils import update_config, upload_workflow_to_s3
+from tibanna.utils import create_jobid
+import boto3
 
 
 def test_update_config(run_task_awsem_event_data):
@@ -48,3 +50,23 @@ def test_update_config5(run_task_awsem_event_omit_fields2):
     assert config['EBS_optimized'] is True
     assert config['ebs_size'] >= 10
     assert config['shutdown_min'] == "now"
+
+
+def test_upload_workflow_to_s3(run_task_awsem_event_cwl_upload):
+    jobid = create_jobid()
+    args = run_task_awsem_event_cwl_upload['args']
+    cfg = run_task_awsem_event_cwl_upload['config']
+    url = upload_workflow_to_s3(args, cfg, jobid)
+    s3 = boto3.client('s3')
+    res1 = s3.get_object(Bucket=cfg['log_bucket'], Key=jobid + '.workflow/main.cwl')
+    res2 = s3.get_object(Bucket=cfg['log_bucket'], Key=jobid + '.workflow/child1.cwl')
+    res3 = s3.get_object(Bucket=cfg['log_bucket'], Key=jobid + '.workflow/child2.cwl')
+    assert res1
+    assert res2
+    assert res3
+    assert url == 's3://tibanna-output/' + jobid + '.workflow/'
+    # clean up afterwards
+    s3.delete_objects(Bucket=cfg['log_bucket'],
+                      Delete={'Objects': [{'Key': jobid + '.workflow/main.cwl'},
+                                          {'Key': jobid + '.workflow/child1.cwl'},
+                                          {'Key': jobid + '.workflow/child2.cwl'}]})