+++ tibanna/core.py
@@ -32,6 +32,9 @@ from .utils import (
     printlog,
     create_jobid,
 )
+from .ec2_utils import (
+    upload_workflow_to_s3
+)
 # from botocore.errorfactory import ExecutionAlreadyExists
 from .iam_utils import (
     create_tibanna_iam,
@@ -136,6 +139,17 @@ class API(object):
         data[_tibanna]['url'] = url
         # add jobid
         data['jobid'] = jobid
+        if 'args' in data:  # unicorn-only
+            args = data['args']
+            cfg = data['config']
+            if ('cwl_directory_local' in args and args['cwl_directory_local']) or \
+                    ('wdl_directory_local' in args and args['wdl_directory_local']):
+                url = upload_workflow_to_s3(args, cfg, jobid)
+                if 'language' in args and args['language'] == 'wdl':
+                    args['wdl_directory_url'] = url
+                else:
+                    args['cwl_directory_url'] = url
+        # submit job as an execution
         aws_input = json.dumps(data)
         print("about to start run %s" % run_name)
         # trigger the step function to run
@@ -163,7 +177,7 @@ class API(object):
                     'home?region=%s#dashboards:name=awsem-%s' % (AWS_REGION, jobid)
                 print("Cloudwatch Dashboard = %s" % cw_db_url)
             if open_browser and shutil.which('open') is not None:
-                subprocess.call('open %s' % data[_tibanna]['url'])
+                subprocess.call(["open", data[_tibanna]['url']])
         return data
 
     def add_to_dydb(self, awsem_job_id, execution_name, sfn, logbucket):